% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhdi.R
\name{myhdi}
\alias{myhdi}
\title{Calculate the Highest Density Interval}
\usage{
myhdi(funame, credwidth = 0.95, tol = 1e-08, ...)
}
\arguments{
\item{funame}{the function name as found in R}

\item{credwidth}{the interval length}

\item{tol}{how accurate (use default)}

\item{...}{extra parameters}
}
\value{
The BCI
}
\description{
Calculate the Highest Density Interval
}
\examples{
myhdi(funame = qbeta , shape1 = 5, shape2 = 7)
}
