% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myvar.R
\name{myvar}
\alias{myvar}
\title{Using compiled C++ to create variance}
\usage{
myvar(x, na_rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{na_rm}{logical for remove NA}
}
\value{
the variance
}
\description{
Using compiled C++ to create variance
}
\examples{
myvar(x = rnorm(30))
}
